LayoutTests/fast/filesystem/entry-points-missing-arguments-expected.txt

Ensures that we get exceptions when we do not include required arguments for webkitRequestFileSystem and webkitResolveLocalFileSystemURL methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS webkitResolveLocalFileSystemURL() threw exception TypeError: Not enough arguments.
PASS webkitResolveLocalFileSystemURL("") threw exception TypeError: Not enough arguments.
PASS webkitRequestFileSystem() threw exception TypeError: Not enough arguments.
PASS webkitRequestFileSystem(TEMPORARY) threw exception TypeError: Not enough arguments.
PASS webkitRequestFileSystem(TEMPORARY, 1) threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
